LPS-147552 - Change dataType value from double to decimal

+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-web/src/main/resources/META-INF/resources/js/custom_fields.js
@@ -1440,7 +1440,7 @@ AUI.add(
 		const DDMDecimalField = A.Component.create({
 			ATTRS: {
 				dataType: {
-					value: 'double',
+					value: 'decimal',
 				},
 
 				fieldNamespace: {
